Allow synchronize to use ControlPath on controller
This allows passing a control_path to an ssh command used by rsync as
part of a synchronize task.

Previously, ControlPersist was disabled by hardcodoing ControlPath=none
(actually, it used the short arg '-S none'). #8473 was the original
issue that lead to this behavior, introduced in [1].
There, rsync was hanging waiting for the ansible ControlMaster session
to timeout. But, the issue only says that ControlPersist was enabled,
without specifying what ControlMaster and ControlPath settings were
getting passed on to rsync for use on its own ssh connection.

Rsync DOES support using ControlPersist/ControlPath (see [2] and [3]).
So, if a user specifies a custom control_path, pass that in whenever the
task requests use_ssh_args, and rsync will be running on the ansible
controller, not on a delegated host. Passing on the control_path does
not make sense on a delegated host.

This does not pass on the hashed ControlPath generated by the ssh
connection plugin. If that is needed, then the ControlPath generation
logic should be extracted into a utility shared by the ssh connection
plugin and the synchronize action plugin (and anything else that needs
the ControlPath on the controller).

[1] e2b8f05b7979616d799914b4dd9839e70082d508 See also:
https://github.com/ansible/ansible-modules-core/commit/989b14de80cd6e41820f2cb98296463811203a78

[2] https://unix.stackexchange.com/questions/50508/reusing-ssh-session-for-repeated-rsync-commands
[3] https://unix.stackexchange.com/questions/48632/cant-share-an-ssh-connection-with-rsync